Datasource
Videogames Data Source.

Schema Explanation:

Rank: This column lists the rank of the game based on its global sales, with 1 being the highest-selling game.

Name: This column contains the title of each game.

Platform: This column specifies the gaming console or platform on which the game was released (e.g., Wii, NES, GB).

Year: This column shows the year in which the game was released.

Genre: This column indicates the genre of the game (e.g., Sports, Platform, Racing).

Publisher: This column lists the company that published the game. In this case, all games are published by Nintendo.

NA_Sales: This column represents the total sales of the game in North America (in millions of units).

EU_Sales: This column shows the total sales of the game in Europe (in millions of units).

JP_Sales: This column indicates the total sales of the game in Japan (in millions of units).

Other_Sales: This column represents the total sales of the game in other regions not specifically listed (in millions of units).

Global_Sales: This column provides the total global sales of the game, which is the sum of NA_Sales, EU_Sales, JP_Sales, and Other_Sales (in millions of units).